requirements/base.txt
Django==3.0.3
django-braces==1.11.0
django-environ==0.4.3
django-model-utils==3.0.0
Pillow==4.1.1
psycopg2==2.7.1
stripe==1.65.0
python-dotenv
psycopg2-binary
Faker==4.0.1
idna==2.9
olefile==0.46
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
stripe==1.65.0
text-unidecode==1.3
urllib3==1.25.8